+++ src/BitChopp.Main/Views/QRCodeWindow.axaml.cs
@@ -60,6 +60,7 @@ public partial class QRCodeWindow : KioskBaseWindow
     {
         try
         {
+            Console.WriteLine("Listening to websocket");
             await _webSocket.ConnectAsync(_wsUrl, CancellationToken.None);
             StartListening();
         }
@@ -77,34 +78,28 @@ public partial class QRCodeWindow : KioskBaseWindow
             var result = await _webSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
             if (result.MessageType != WebSocketMessageType.Text)
             {
-                if (result.MessageType != WebSocketMessageType.Text)
-                {
-                    Console.WriteLine("Received non-text WebSocket message.");
-                    continue;
-                }
-
-                var message = Encoding.UTF8.GetString(buffer, 0, result.Count).Split('-');
-
-                var pin = int.Parse(message[0]);
-                // var duration = int.Parse(message[1]);
-
-                // TODO: Check PIN id; message = pinId-duration
-                if (pin != _pinId)
-                {
-                    if (pin != _pinId)
-                    {
-                        Console.Error.WriteLine("Invalid PIN. Someone probably paid old invoice");
-                        return;
-                    }
-
-                    WebSocketResult = "Paid";
-
-                    _ = _webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "", CancellationToken.None);
-
-                    Close();
-                    return;
-                }
+                Console.WriteLine("Received non-text WebSocket message.");
+                continue;
             }
+
+            var message = Encoding.UTF8.GetString(buffer, 0, result.Count).Split('-');
+
+            var pin = int.Parse(message[0]);
+            // var duration = int.Parse(message[1]);
+
+            // TODO: Check PIN id; message = pinId-duration
+            if (pin != _pinId)
+            {
+                Console.Error.WriteLine("Invalid PIN. Someone probably paid old invoice");
+                return;
+            }
+
+            WebSocketResult = "Paid";
+
+            _ = _webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "", CancellationToken.None);
+
+            Close();
+            return;
         }
     }
 }
